refactor(services): merging ui and feature libs under a same one
