+++ libs/pages/services/README.md
@@ -1,0 +1,7 @@
+# pages-services
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test pages-services` to execute the unit tests via [Jest](https://jestjs.io).
