--- libs/pages/services/feature/project.json
@@ -1,22 +1,0 @@
-{
-  "sourceRoot": "libs/pages/services/feature/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/pages/services/feature/**/*.{ts,tsx,js,jsx}"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/pages/services/feature"],
-      "options": {
-        "jestConfig": "libs/pages/services/feature/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  }
-}
