--- libs/pages/services/feature/src/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/services-page'
-export * from './lib/components/deployments/deployments'
-export * from './lib/components/general/general'
-export * from './lib/router/router'
