--- libs/pages/services/feature/src/lib/services-page.spec.tsx
@@ -1,9 +1,0 @@
-import { render } from '__tests__/utils/setup-jest'
-import ServicesPage from './services-page'
-
-describe('ServicesPage', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<ServicesPage />)
-    expect(baseElement).toBeTruthy()
-  })
-})
