+++ libs/pages/services/jest.config.ts
@@ -1,10 +1,10 @@
 /* eslint-disable */
 export default {
-  displayName: 'pages-applications-ui',
-  preset: '../../../../jest.preset.js',
+  displayName: 'pages-services',
+  preset: '../../../jest.preset.js',
   transform: {
     '^.+\\.[tj]sx?$': 'babel-jest',
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/libs/pages/services/ui',
+  coverageDirectory: '../../../coverage/libs/pages/services',
 }
