+++ libs/pages/services/project.json
@@ -1,5 +1,6 @@
 {
-  "sourceRoot": "libs/pages/services/ui/src",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/pages/services/src",
   "projectType": "library",
   "tags": [],
   "targets": {
@@ -7,14 +8,14 @@
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/pages/services/ui/**/*.{ts,tsx,js,jsx}"]
+        "lintFilePatterns": ["libs/pages/services/**/*.{ts,tsx,js,jsx}"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/pages/services/ui"],
+      "outputs": ["coverage/libs/pages/services"],
       "options": {
-        "jestConfig": "libs/pages/services/ui/jest.config.ts",
+        "jestConfig": "libs/pages/services/jest.config.ts",
         "passWithNoTests": true
       }
     }
