+++ libs/pages/services/src/lib/feature/deployments/deployments-feature.spec.tsx
@@ -1,0 +1,9 @@
+import { render } from '__tests__/utils/setup-jest'
+import DeploymentsFeature from './deployments-feature'
+
+describe('Deployments', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<DeploymentsFeature />)
+    expect(baseElement).toBeTruthy()
+  })
+})
