--- libs/pages/services/ui/src/index.ts
@@ -1,4 +1,0 @@
-export * from './lib/components/table-row-services/table-row-services'
-export * from './lib/components/deployments-page/deployments-page'
-export * from './lib/components/general/general'
-export * from './lib/components/container/container'
