+++ tsconfig.base.json
@@ -25,16 +25,15 @@
       "@console/pages/application/ui": ["libs/pages/application/ui/src/index.ts"],
       "@console/pages/environments/feature": ["libs/pages/environments/feature/src/index.ts"],
       "@console/pages/environments/ui": ["libs/pages/environments/ui/src/index.ts"],
-      "@console/pages/login": ["libs/pages/login/src/index.ts"],
       "@console/pages/layout": ["libs/pages/layout/src/index.ts"],
+      "@console/pages/login": ["libs/pages/login/src/index.ts"],
       "@console/pages/logs/infra/feature": ["libs/pages/logs/infra/feature/src/index.ts"],
       "@console/pages/logs/infra/ui": ["libs/pages/logs/infra/ui/src/index.ts"],
       "@console/pages/onboarding/feature": ["libs/pages/onboarding/feature/src/index.ts"],
       "@console/pages/onboarding/ui": ["libs/pages/onboarding/ui/src/index.ts"],
       "@console/pages/overview/feature": ["libs/pages/overview/feature/src/index.ts"],
       "@console/pages/overview/ui": ["libs/pages/overview/ui/src/index.ts"],
-      "@console/pages/services/feature": ["libs/pages/services/feature/src/index.ts"],
-      "@console/pages/services/ui": ["libs/pages/services/ui/src/index.ts"],
+      "@console/pages/services": ["libs/pages/services/src/index.ts"],
       "@console/pages/settings/feature": ["libs/pages/settings/feature/src/index.ts"],
       "@console/pages/settings/ui": ["libs/pages/settings/ui/src/index.ts"],
       "@console/shared/auth": ["libs/shared/auth/src/index.ts"],
