+++ workspace.json
@@ -22,8 +22,7 @@
     "pages-onboarding-ui": "libs/pages/onboarding/ui",
     "pages-overview-feature": "libs/pages/overview/feature",
     "pages-overview-ui": "libs/pages/overview/ui",
-    "pages-services-feature": "libs/pages/services/feature",
-    "pages-services-ui": "libs/pages/services/ui",
+    "pages-services": "libs/pages/services",
     "shared-auth": "libs/shared/auth",
     "shared-enums": "libs/shared/enums",
     "shared-interfaces": "libs/shared/interfaces",
